renaming Setup -> Startup
renaming Setup/Cleanup function pair to Startup/Cleanup.
This function pair is meant to make it easier (compared to using
class Ctor/Dtor pair) to aggregate classes on top of each other
for better cache locality.

other candidate names for the function pair:
Create/Delete: usually implies memory allocation, but there is none.
Init/Shutdown: verbs for large systems or managers.
Begin/End: misleading for iterator classes.

Hopefully Startup/Cleanup is enough to annotate the begin and end of
the lifetime of an object sitting on top of allocated memory. In the
codebase, Ctor/Dtor only performs lightweight checks and do not define
the object lifetime.

Why is naming so difficult.