RenderBase, single threaded single device API
BeginFrame, EndFrame defines frame boundaries.
Add three vulkan swapchain related functions
- GetSwapChainTextureFormat
- GetSwapChainRenderPass
- GetSwapChainFrameBuffer

The vulkan backend defines a final render pass where we blit
the final output to the framebuffer corresponding to the current
swapchain image. Frames in flight are not possible in OpenGL and
therefore must be abstracted by the backend.